# Import slow modules lazily