@nateprewitt 2017-05-25
So this is really interesting. Kenneth and I identified that we're spending a lot of time importing the other night while pairing but didn't catch that much of it was on these two modules.

I'm onboard with this if the time numbers are correct. Is there a reason you used two different commands?