@nonylene 2017-05-25
Thanks!

> I'm onboard with this if the time numbers are correct. Is there a reason you used two different commands?

`pipenv` runs pipenv installed by pip . `puthon3 -m pipenv` runs my modified code.