@nateprewitt 2017-05-25
That's what I figured, I'm just curious if there's any variation in overhead by running one way or another. It's purely curiousity at this point, I can poke around.